+++ nginx/config/entrypoint.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 # Wazuh App Copyright (C) 2018 Wazuh Inc. (License GPLv2)
 
+
 set -e
 
 # Generating certificates.
@@ -37,7 +38,6 @@ if [ ! -f /etc/nginx/conf.d/kibana.htpasswd ]; then
 else
   echo "Kibana credentials already configured"
 fi
-
 
 
 if [ "x${NGINX_PORT}" = "x" ]; then